% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_eqtl_data}
\alias{fetch_eqtl_data}
\title{Cross-population eQTL and fine-mapping API}
\usage{
fetch_eqtl_data(gene)
}
\description{
Cross-population eQTL and fine-mapping API
}
\examples{
eqtls = fetch_eqtl_data('ORMDL3')
eqtls = fetch_eqtl_data('BRCA1')
}

