% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_eqtl_locus}
\alias{plot_eqtl_locus}
\title{Plot both multipopulation eQTL data and fine-mapping eQTL data for a specific population.}
\usage{
plot_eqtl_locus(obj, populations = c("all"), discovery = "CEU",
  start = -1, end = -1)
}
\description{
Plot both multipopulation eQTL data and fine-mapping eQTL data for a specific population.
}
\examples{
eqtls = fetch_eqtl_data('SP1')
plot_eqtl_locus(eqtls)

eqtls = fetch_eqtl_data('ORMDL3')
plot_eqtl_locus(eqtls, populations=c('CHB','JPT'), discovery='JPT')
}

