% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_finemapped_eqtls}
\alias{plot_finemapped_eqtls}
\title{Plot fine-mapping eQTL data for a specific discovery population}
\usage{
plot_finemapped_eqtls(obj, discovery = "CEU", start = -1, end = -1)
}
\description{
Plot fine-mapping eQTL data for a specific discovery population
}
\examples{
eqtls = fetch_eqtl_data('ORMDL3')
plot_finemapped_eqtls(eqtls, 'JPT')
plot_finemapped_eqtls(eqtls, 'JPT', start = 38000000, 38050000)
}

