% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_multipopulation_eqtls}
\alias{plot_multipopulation_eqtls}
\title{Plot eQTL data for each population}
\usage{
plot_multipopulation_eqtls(obj, populations = c("all"), start = -1,
  end = -1)
}
\description{
Plot eQTL data for each population
}
\examples{
eqtls = fetch_eqtl_data('BRCA1')
plot_multipopulation_eqtls(eqtls, populations='CEU')
plot_multipopulation_eqtls(eqtls, populations=c('CEU','JPT'))
plot_multipopulation_eqtls(eqtls, start = 41000000, end = 41500000)
}

